% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambr_grid_plot.R
\name{ambr_grid_plot}
\alias{ambr_grid_plot}
\title{Create a grid plot for Ambr data}
\usage{
ambr_grid_plot(data)
}
\arguments{
\item{data}{The dataframe with Ambr data}
}
\value{
A ggplot object with the grid plot
}
\description{
This function takes a dataframe with Ambr data and returns a ggplot object with a grid plot of the data
}
\examples{
\dontrun{
data <- data.frame(time_from_inoculation = runif(100),
                   value = runif(100), vessel_id = sample(letters, 100, replace = TRUE))
ambr_grid_plot(data)
}
}
