% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ambr_data.R
\name{ambr_map}
\alias{ambr_map}
\title{Ambr Map}
\usage{
ambr_map()
}
\value{
A dataframe with the map
}
\description{
Gives a map of the ambr with culture stations, vessels ids and vessel numbers according to standards
}
