% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_map.R
\name{append_map}
\alias{append_map}
\title{Appends missing mapping values to the dataframe}
\usage{
append_map(data, append_by)
}
\arguments{
\item{data}{The dataframe to which the ambr map will be appended.}

\item{append_by}{The columns on which to join the dataframe and the ambr map.}
}
\value{
A dataframe with the ambr map appended.
}
\description{
This function joins the input dataframe with the ambr map on
the specified columns.
}
\examples{
# Example usage
df <- data.frame(culture_station = c("CS1", "CS2"), vessel_number = c(1, 2))
append_map(df, c("culture_station", "vessel_number"))

}
