% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ambr_audit_data.R
\name{import_ambr_audit_data}
\alias{import_ambr_audit_data}
\title{Import Ambr audit data}
\usage{
import_ambr_audit_data(path, schema = FALSE)
}
\arguments{
\item{path}{The directory path from which to import the CSV files}

\item{schema}{An optional parameter to specify a schema to be applied to the data}
}
\value{
A list of dataframes, each representing a parameter in the Ambr audit data
}
\description{
This function imports Ambr audit data from CSV files in a specified directory
}
\examples{
# Example usage
import_ambr_audit_data(utils::choose.dir(), schema = FALSE)

}
