% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_sheet_name.R
\name{sanitize_sheet_name}
\alias{sanitize_sheet_name}
\title{Sanitize sheet names for Excel}
\usage{
sanitize_sheet_name(name)
}
\arguments{
\item{name}{The original sheet name}
}
\value{
The sanitized sheet name
}
\description{
This function cleans up sheet names for Excel by replacing special characters and spaces with underscores,
and removing leading or trailing underscores and square brackets
}
\examples{
sanitize_sheet_name("Sheet [1]")

}
