% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scientific_10.R
\name{scientific_10}
\alias{scientific_10}
\title{Function to display scientific notation on the y axis}
\usage{
scientific_10(x)
}
\arguments{
\item{x}{The number to convert.}
}
\value{
A string representing the number in scientific notation.
}
\description{
This function converts a number into a string
that represents the number in scientific notation.
}
\examples{
scientific_10(10000)

}
