% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ambr_data.R
\name{append_map}
\alias{append_map}
\title{Append map}
\usage{
append_map(data, append_by)
}
\arguments{
\item{data}{the data to append the map to}

\item{append_by}{the .data$parameter, generally c("culture_station", "vessel_number")}
}
\value{
A stitch of the dataframe with the map of the positions
}
\description{
Appends missing mapping values to the dataframe
}
