% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ambr_data.R
\name{import_ambr_audit_data}
\alias{import_ambr_audit_data}
\title{Main function of the package, imports all ambr data in a frame}
\usage{
import_ambr_audit_data(path = "AuditData", schema = F)
}
\arguments{
\item{path}{the path of the AuditData folder}

\item{schema}{a schema in csv format with clone and generation as additional parameters}
}
\value{
A list of dataframe with all the parameters for the ambr
}
\description{
Main function of the package, imports all ambr data in a frame
}
