% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_transfection_time.R
\name{import_transfection_time}
\alias{import_transfection_time}
\title{Import and clean transfection time data}
\usage{
import_transfection_time(path = utils::choose.dir())
}
\arguments{
\item{path}{The directory path from which to import the CSV files}
}
\value{
A dataframe with the cleaned transfection time data
}
\description{
This function imports and cleans transfection time data from a CSV file
}
\examples{
import_transfection_time()

}
