% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ambr_data.R
\name{import_transfection_time}
\alias{import_transfection_time}
\title{Function to import transfection time from process audit}
\usage{
import_transfection_time(path = "AuditData/ProcessAudit.csv")
}
\arguments{
\item{path}{the location of ProcessAudit.csv}
}
\value{
a dataframe with transfection timepoints in standard format
}
\description{
Function to import transfection time from process audit
}
